+++ DnDGenerator/src/main/java/com/revature/data/Dao.java
@@ -2,6 +2,8 @@ package com.revature.data;
 
 import java.sql.Connection;
 
+import com.revature.utilities.ConnectionHandler;
+
 public class Dao {
 	
 	Connection getConnection() {
